+++ src/CONST.js
@@ -3,7 +3,7 @@
  * @Author: OBKoro1
  * @Created_time: 2019-05-14 15:41:59
  * @LastEditors: OBKoro1
- * @LastEditTime: 2019-05-27 16:24:42
+ * @LastEditTime: 2019-08-08 13:45:55
  * @Description: 常量文件
  */
 
@@ -14,7 +14,6 @@ const middleAnnotation = {
     vb: "' ",
     html: '* ',
     shellscript: ' # '
-    // userLanguage: '' // 用户自定义的单行
 }
 
 module.exports = {